Add per_page param

+++ app/Repositories/CemeteryRepository.php
@@ -26,7 +26,7 @@ Class CemeteryRepository implements RepositoryInterface, CemeteryRepositoryInter
                      ->orWhere('name', 'like' , '%'.$search.'%');
         });
 
-        return $this->model->paginate(request('per_page') ?? Cemetery::POSTS_PER_PAGE);
+        return $this->model->paginate((int)request('per_page') ?? Cemetery::POSTS_PER_PAGE);
     }
     
     public function find($id)
